sad file moving noises
