+++ src/main/cpp/subsystems/RobotContainer.cpp
@@ -2,8 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "RobotContainer.h"
-#include <iostream>
+#include "subsystems/RobotContainer.h"
 
 RobotContainer::RobotContainer() : m_autonomousCommand(&m_subsystem) {
   // Initialize all of your commands and subsystems here
@@ -31,13 +30,7 @@ void RobotContainer::UpdateControls() {
   } else {
     _ledController.Update();
   }
-  // if (_controlBoard.GetFlywheelDesired()) {
-  //   frc2::CommandScheduler::GetInstance().Schedule(true, &_shootCommand);
-  //   _ledController.UpdateShooting();
-  // }
-  // else {
-  //   _ledController.Update();
-  // }
+
   _controlBoard.ReadControls();
 }
