+++ src/main/include/subsystems/RobotContainer.h
@@ -8,7 +8,7 @@
 
 #include "commands/ExampleCommand.h"
 #include "subsystems/ExampleSubsystem.h"
-#include "ControlBoard.h"
+#include "subsystems/ControlBoard.h"
 
 #include "commands/DrivetrainCommand.h"
 #include "commands/ShootCommand.h"
